Porting recent internal changes to public version.
Summary of changes:
1) Switching the default rnn model to be bidirectional.
2) Introducing new evaluation metrics. Sometimes there are multiple library entries for a given molecule.
Previously, all of our metrics were in terms of the highest-ranked entry for a given molecule. Now, we also compute
metrics in terms of the average rank of all of the library entries for a given molecule.
These new metrics have analogous names as the original ones, but are prefixed with 'avg-rank-.'